+++ tests/stats.js
@@ -3,10 +3,37 @@ const assert = require('assert');
 const getStats = require('../lib/twilio/rtc/stats');
 const standardPayload = require('./payloads/rtcstatsreport.json');
 const edgePayload = require('./payloads/rtcstatsreport-edge.json');
+const ffPayload = require('./payloads/rtcstatsreport-ff.json');
 const legacyPayload = require('./payloads/rtcstatsresponse.json');
 const MockRTCStatsReport = require('../lib/twilio/rtc/mockrtcstatsreport');
 
 describe('getStats', () => {
+  context('In Firefox', () => {
+    let stats;
+
+    before(() => {
+      const statsReport = MockRTCStatsReport.fromArray(ffPayload);
+      const peerConnection = {
+        getStats() { return Promise.resolve(statsReport); }
+      };
+
+      return getStats(peerConnection).then(_stats => { stats = _stats; });
+    });
+
+    it('should correctly transform a FF RTCStatsReport', () => {
+      assert.deepEqual(stats, {
+        bytesReceived: 143792,
+        bytesSent: 144996,
+        jitter: 3,
+        packetsLost: 0,
+        packetsReceived: 836,
+        packetsSent: 843,
+        rtt: 81,
+        timestamp: 1583785892295
+      });
+    });
+  });
+
   context('In standard browsers', () => {
     let stats;
 
@@ -26,7 +53,6 @@ describe('getStats', () => {
         codecName: 'PCMU',
         timestamp: 1492027598825.6,
         rtt: 86,
-        rttSeconds: 0.086,
         jitter: 3,
         packetsLost: 41,
         packetsReceived: 61687,
@@ -58,7 +84,6 @@ describe('getStats', () => {
         codecName: 'PCMU',
         timestamp: 1492027598825.6,
         rtt: 86,
-        rttSeconds: 0.086,
         jitter: 3,
         packetsLost: 41,
         packetsReceived: 61687,
@@ -90,7 +115,6 @@ describe('getStats', () => {
         codecName: 'opus',
         timestamp: 1492027598825.6,
         rtt: 86,
-        rttSeconds: 0.086,
         jitter: 3,
         packetsLost: 41,
         packetsReceived: 61687,
@@ -122,7 +146,6 @@ describe('getStats', () => {
         codecName: 'PCMU',
         timestamp: 1492472862784,
         rtt: 85,
-        rttSeconds: 0.085,
         jitter: 4,
         packetsLost: 0,
         packetsReceived: 24776,